requirements.txt
astroid==2.1.0
autopep8==1.4.3
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.4.0
django-url-or-relative-url-field==0.1.1
djangorestframework==3.9.0
entrypoints==0.3
flake8==3.7.4
flake8-builtins==1.4.1
flake8-docstrings==1.3.0
flake8-import-order==0.18
flake8-polyfill==1.0.2
flake8-quotes==1.0.0
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pathlib==1.0.1
Pillow==5.4.1
pycodestyle==2.5.0
pydocstyle==3.0.0
pyflakes==2.1.0
pylint==2.2.2
python-decouple==3.1
pytz==2018.9
six==1.12.0
snowballstemmer==1.2.1
Unipath==1.1
wrapt==1.11.0
